+++ resources/migrations/20191125085721-generic-outbox.up.sql
@@ -1,0 +1,4 @@
+CREATE TABLE outbox (
+  id serial NOT NULL PRIMARY KEY,
+  outboxData jsonb NOT NULL
+)